Created Models and Services, started to clean up code
